// src/api/generated/models/InputValueType.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum InputValueType {
    INDICATOR = 'Indicator',
    NET_TO_GROSS = 'NetToGross',
    VARIOGRAM_FAMILY_FILTER = 'VariogramFamilyFilter',
    ATTRIBUTE_NAME = 'AttributeName',
    ARCHEL = 'Archel',
}